+++ silo-core/contracts/incentives/interfaces/ISiloIncentivesController.sol
@@ -64,7 +64,7 @@ interface ISiloIncentivesController is IDistributionManager {
      * @param _amount The amount of rewards to distribute
      * @return programId The id of the created or existing program, or bytes32(0) if _amount is 0
      */
-    function immediateDistribution(address _tokenToDistribute, uint104 _amount) external returns (bytes32 programId);
+    function immediateDistribution(address _tokenToDistribute, uint256 _amount) external returns (bytes32 programId);
 
     /// @dev It will transfer all the reward token balance to the owner.
     /// @param _rewardToken The reward token to rescue
